Finished unit testing models
Fixes #27: DeriveNeume test still has 3 errors, but they are known about and will be fixed when issue #30 is completed (converting lookup to a binary tree search). The problem is that both a punctum and a virga are neumes, but each are one note. So look up stops at first match. When doing the binary tree lookup, both a punctum and a virga will be in the 'bin' of [0] and will have to discriminate the two by the head of the neume component.