Added group functionality
Individual glyphs can now be grouped together in the view and manipulated together.
Also added references to neumes on the view objects so we can access underlying model information on view interactions.
Did a quick test: when a fabric neume group is clicked the neume name is printed and the colour changes.