Attach staffRef to canvas objects
in order to know what the element is a part of. Otherwise you have an ID and you would have to search through all the staves to find where it is to remove it from the staff. It's either this or we have some massive dictionary on the page that maps element id's to staves.