GUI work
Added a progress bar to the layout, which may come in handy for operations on large selections of elements. 25% of the progress bar width can be server side, 75% can be client side.
Cleaned up sidebar menu. Clicking the insert button brings up options for inserting punctum or divisions and the punctum doesn't immediately follow the pointer anymore.