Manual width height parameters
The width and height of the original document can now be sent as
parameters to Neon, instead of gathering the dimension information from
the background image. This means that Neon can now be served smaller
dimension images for the background image, and still output correct
bounding boxes etc. to MEI.