Pluggable drawing libraries
Now have pluggable drawing code for both the liber and salzinnes
manuscripts (or any other that need to be defined).
The desired drawing library name is specified as a Neon parameter. The
drawing library name is specified as a parameter to the bash compilation
script, so the right library code is compiled in.