St. Gallen neume drawing
Created custom SVG glyphs for St. Gallen manuscript in adobe
illustrator and exported each to SVG, which is then parsed by Neon.js as
independent glyphs. Created Neon.js drawing code for the fundamental
glyphs in St. Gallen and it displays. Need to think about how to set the
global scale of glyphs. With staffed neumes, the global scale could be
set by comparing the size of the SVG to the distance between staff
lines.