Writing liquescent neume mei
Hold off on reading and writing liquescent neumes in the MEI file until
an encoding format has been determined. So, the interface can draw
augmented and diminished liquescent neumes but can not save or read them
from the MEI file. Therefore, upon refreshing the page, the liquescent
drawings of the neume will disappear and the normal drawings will be
shown.