Fixed #45: We can insert staves!
The size of the new staff is that of the largest staff on the page.  Also, the user will have to define which number/order this staff is (default is the "next", but if you are inserting staves at the beginning of the page or in the middle, you have to set the order manually with the slider.  All other staves will automatically have their order number changed accordingly.)