WIP: debug_neon and demo config settings are independent
This commit allows for one to develop the demo where both staffless and square files can be displayed and edited, or to develop a production server. The demo uses the neon_square/neon_staffless template; production uses plain neon templates. If NEON_DEBUG in conf.py is turned on, Neon will serve the _devel files with uncompiled javascript; if NEON_DEBUG is turned off, Neon will serve the _prod files with compiled javascript.

server.py and interface.py changes reflect this: different filepaths are needed depending on whether the squarenote and cheironomic files are in the data folder.

Template file changes reflect the different files needed to load these.

This is an intermediary commit; this process will be refined later.