Fixed: for Salzinnes curved staff lines
Solution: we have to use separate systems to fit the curved staff lines, but it is now allowed that a system have no clef. Neon.js will find a clef from previous systems, and use that clef to calculate the pitches in this system. Besides, the custos is not created within an interrupted system (if a system has no clef, Neon.js will not create a custos on the previous system).