Fixed #131: Can remove lone clefs
Clefs that are the only element on a staff can now be removed. This is important especially when an entire system has no elements on it, which is a common occurrence within the salzinnes. If the user inserted their first clef in the wrong position, they can now edit it.