#124: Preliminary Fix
You can now deselect groups of objects. The issue before was that the handleEventObjectModified class would try to switch on the eleRef.constructor, but with groups, they don’t even have an eleRef property so it would be undefined. I just inserted a simple check to see if the selection has an eleRef property, and to only execute the code if it does.

A new issue appeared now that deselection is possible though that the snapping is really off on group movement, so something needs to be done about that.