Fixed #140 - Revert now works
Okay this has been a recurring issue so I’m not sure how the original programmers dealt with this. Any time we have to deal with filenames, it always seems to be missing an .mei at the end for it to be a valid path. So I added it into the code for revert and the functionality works.