#175 - Proper Custos Inserting
Custos now render properly when inserted. The issue was that when instantiating a custos, it was not possible to pass through the pname / octave arguments as in the punctum insertion, those are done after the clef check. I just had to rearrange the code so that the check still happens properly, but also the arguments can be passed.