Working #119: Can now undo
Undo functionality now works. By saving various stages of the working document to an “undo” folder, you can undo back to earlier versions of the document.

Now we need to find a way to clear all the created documents and when to do so. We would like it to be during the initial loading of the neon page, but this is a problem since some functionality depends on refreshing the page.

To fix this I’m going to work on issue #182 and try and find a way to refresh the canvas without refreshing the page, and then return to the undo functionality.