#177 - Finalized System Merges
Systems can now be merged easily with the merge system option. The two caveats are, one of the systems must be empty (Always is from the OMR process), and the page must be refreshed for the elements to appear on the new system. This is because of the way rendering acts with inserting a new clef onto a system.