Fixed #191: default document type squarenote
Since we are not including cheironomic functionality for quite some time, we set the default document type to be squarenote.
However, I still updated the commented out code in demo.html so that undo does not show up in the dropdown menu so that once someone does begin to work on cheironomic again, it will be functional.