postponing #196: Duplication
Cannot seem to find a way to make duplication work without having to refresh the page in order for the duplicate image to render. This would not be intuitive to the user because upon refresh, all the undo files are lost. So if they mistakingly duplicate many neumes, they would be unable to undo their action effectively. Therefore we are postponing until after the first release.